casting point coordinates to float for svd

+++ src/svd.cpp
@@ -10,15 +10,15 @@ std::pair<Eigen::JacobiSVD<Eigen::MatrixXf>, Eigen::MatrixXf> svd::compute(
     Eigen::MatrixXf vectors(points.size(), R);
     int row = 0;
     for (auto point : points) {
-        vectors(row, xCol) = point.m_xyz[0];
-        vectors(row, yCol) = point.m_xyz[1];
-        vectors(row, zCol) = point.m_xyz[2];
+        vectors(row, xCol) = (float)point.m_xyz[0];
+        vectors(row, yCol) = (float)point.m_xyz[1];
+        vectors(row, zCol) = (float)point.m_xyz[2];
         row++;
     }
     /**   translate points to direction vectors */
-    vectors.col(xCol).array() -= centroid.m_xyz[0];
-    vectors.col(yCol).array() -= centroid.m_xyz[1];
-    vectors.col(zCol).array() -= centroid.m_xyz[2];
+    vectors.col(xCol).array() -= (float)centroid.m_xyz[0];
+    vectors.col(yCol).array() -= (float)centroid.m_xyz[1];
+    vectors.col(zCol).array() -= (float)centroid.m_xyz[2];
 
     /** compute svd */
     int setting = Eigen::ComputeThinU | Eigen::ComputeThinV;
